Get ProcessOSM compiling with jts geoms

+++ src/analytics/src/main/scala/osmesa/analytics/Footprints.scala
@@ -666,7 +666,7 @@ object Footprints extends Logging {
             if (value > 0) {
               features.append(
                 PointFeature(Point(rasterExtent.gridToMap(c, r)),
-                             (index.toIndex(SpatialKey(c, r)), value)))
+                             (index.toIndex(SpatialKey(c, r)).toLong, value)))
             }
           }
 
@@ -693,7 +693,7 @@ object Footprints extends Logging {
               if (value > 0) {
                 features.append(
                   PointFeature(Point(rasterExtent.gridToMap(c, r)),
-                               (index.toIndex(SpatialKey(c, r)), value)))
+                               (index.toIndex(SpatialKey(c, r)).toLong, value)))
               }
             }
